refactor(config): update .env loading and fix exception logging
Updates environment variable loading to be compatible with Pydantic v2. The _env_file attribute in BaseSettings is replaced by passing the nv_file argument directly to the settings constructor.

Removes xc_info=False from logger.exception calls. This was a bug, as logger.exception is intended to always log the exception traceback, and this argument was suppressing it.

BREAKING CHANGE: The application no longer loads a .env file by default. The --env-file argument must now be used to specify an environment file.